requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
openai==1.99.1
python-docx==1.1.0 